+++ __tests__/__system__/cli/submit/data-set/__scripts__/command/command_submit_data_set_waitopt.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+dsn=$1
+option=$2
+set -e
+
+bright zos-ftp submit data-set "$dsn" "$option"