+++ __tests__/__system__/cli/submit/local-file/__scripts__/command/command_submit_local_file_waitopt.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+localFile=$1
+option=$2
+set -e
+
+bright zos-ftp submit local-file "$localFile" "$option"