+++ __tests__/__system__/cli/submit/stdin/__scripts__/command/command_submit_stdin_waitopt.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+localFile=$1
+option=$2
+set -e
+set -o pipefail
+
+cat "$localFile" | bright zos-ftp submit stdin "$option"