+++ __tests__/__system__/cli/submit/stdin/cli.submit.stdin.test.ts
@@ -76,7 +76,7 @@ describe("submit job from stdin command", () => {
         expect(result.output.toString()).toContain("rc:");
     });
 
-    it("should be able to submit a job from a local file with invalid wait value", async () => {
+    it("should give a syntax error if the wait value is invalid", async () => {
 
         // download the appropriate JCL content from the data set
         const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
@@ -85,11 +85,9 @@ describe("submit job from stdin command", () => {
         await IO.writeFileAsync(jclFilePath, sleepContent);
         const wait = "3,";
         const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,wait]);
-        expect(result.stderr.toString()).toEqual("");
+        expect(result.output.toString()).toContain("comma-separated numeric values");
+        expect(result.output.toString()).toContain("Syntax error:");
         expect(result.status).toEqual(0);
-        expect(result.output.toString()).toContain("is invalid.");
-        expect(result.output.toString()).toContain("The default value");
-        expect(result.output.toString()).toContain("rc:");
     });
 
     it("should be able to submit a job from a local file but not finished within specified wait option", async () => {
@@ -108,4 +106,34 @@ describe("submit job from stdin command", () => {
         expect(result.output.toString()).toContain("Please using the following command to check its status later:");
     });
 
+    it("should be able to submit a job from a local file with wait-for-output option and get rc successfully", async () => {
+
+        // download the appropriate JCL content from the data set
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option = "--wait-for-output";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_waitopt.sh", testEnvironment, [jclFilePath,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Waiting for job completion.");
+        expect(result.output.toString()).toContain("rc:");
+        expect(result.output.toString()).toContain("finished");
+    });
+
+    it("should be able to submit a job from a local file with wait-for-active option and get rc successfully", async () => {
+
+        // download the appropriate JCL content from the data set
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option = "--wait-for-active";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_waitopt.sh", testEnvironment, [jclFilePath,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).not.toContain("rc:");
+        expect(result.output.toString()).toContain("ACTIVE");
+    });
 });